requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
aiosqlite
python-jose[cryptography]
passlib[bcrypt]
structlog
google-genai
python-dotenv
httpx
tenacity
networkx
numpy
flask-cors
watchfiles
